requirements.txt
pymongo==4.6.1
Werkzeug==2.2.2
flask==2.2.2
functions-framework==3.*
pyyaml== 6.0